requirements.txt
djangorestframework==3.7.7
Django==2.0.1
django-rest-auth==0.9.3
django-allauth==0.36.0
django-filter==1.1.0
Pygments==2.2.0
Markdown==2.6.10
coreapi==2.3.3
psycopg2==2.7.3.2
dj-database-url==0.4.2
gunicorn==19.7.1
whitenoise==3.3.1
